requirements.txt
flask==3.0.0
requests==2.31.0
python-telegram-bot==20.7
psutil==5.9.7
apscheduler==3.10.4
